+++ front-end/src/App.js
@@ -3,14 +3,14 @@ import './App.css';
 import Home from './routes/home/home.component';
 import Header from './Header';
 import Viewall from './Viewall';
-import CourseRatings from './CourseRatings';
-import CourseReviews from './CourseReviews';
-import CourseDetails from './CourseDetails';
+import CourseRatings from './components/course-ratings/CourseRatings';
+import CourseReviews from './components/course-reviews/CourseReviews';
+import CourseDetails from './components/course-details/CourseDetails';
 import Footer from './Footer';
 import BottomNavigationBar from './routes/bottom-navigation-bar/bottom-navigation-bar.component';
-import AddReview from './AddReview';
+import AddReview from './routes/add-review/AddReview';
 import SearchPage from "./routes/search-page/search-page.component";
-import ContactUs from "./routes/contact_us/ContactUs";
+import ContactUs from "./routes/contact-us/ContactUs";
 import Login from './Login';
